edit HttpResponse data return

+++ blog/views.py
@@ -1,9 +1,12 @@
+import json
+
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
-from django.http import JsonResponse, HttpResponse, HttpResponseNotFound, HttpResponseBadRequest, HttpResponseNotAllowed
+from django.http import HttpResponse, HttpResponseNotFound, HttpResponseBadRequest, HttpResponseNotAllowed
 from django.views.decorators.csrf import csrf_exempt
-from rest_framework import permissions
 from rest_framework import viewsets
+from rest_framework.permissions import IsAuthenticated
 
 from blog import models
 from blog.forms import UserForm, UserInfoForm
@@ -13,19 +16,22 @@ from blog.serializers import PostSerializer, UserSerializer
 class UserViewSet(viewsets.ModelViewSet):
     queryset = User.objects.all().order_by('-date_joined')
     serializer_class = UserSerializer
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated]
 
 
 class PostViewSet(viewsets.ModelViewSet):
     queryset = models.Post.objects.all().order_by('date')
     serializer_class = PostSerializer
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated, ]
 
 
+@login_required
 @csrf_exempt
 def user_logout(request):
     logout(request)
-    return HttpResponse("User logged out")
+    data_set = dict(detail="User Logged Out")
+    json_response = json.dumps(data_set)
+    return HttpResponse(json_response)
 
 
 @csrf_exempt
@@ -37,13 +43,22 @@ def user_login(request):
         if user:
             if user.is_active:
                 login(request, user)
-                return HttpResponse("User Logged-in successfully.")
+                user_ = models.User.objects.get(username=username)
+                data_set = dict(username=user_.username, email=user_.email)
+                json_response = json.dumps(data_set)
+                return HttpResponse(json_response)
             else:
-                return HttpResponseNotFound("Account Not Active!")
+                data_set = dict(detail="Account Not Active.")
+                json_response = json.dumps(data_set)
+                return HttpResponseNotFound(json_response)
         else:
-            return HttpResponseNotFound("Invalid login details!")
+            data_set = dict(detail="Invalid Login Details.")
+            json_response = json.dumps(data_set)
+            return HttpResponseNotFound(json_response)
     else:
-        return HttpResponseBadRequest("Bad Request.")
+        data_set = dict(detail="Bad Request.")
+        json_response = json.dumps(data_set)
+        return HttpResponseBadRequest(json_response)
 
 
 @csrf_exempt
@@ -59,16 +74,21 @@ def register(request):
             user_profile.user = user
             user_profile.save()
             registered = True
-            data = "username: " + str(user_form['username'].value()) + \
-                   ", email: " + str(user_form['email'].value()) + \
-                   ", registered: " + str(registered)
-            return HttpResponse(data)
+            data_set = dict(username=str(user_form['username'].value()), email=str(user_form['email'].value()),
+                            registered=registered)
+            json_response = json.dumps(data_set)
+            return HttpResponse(json_response)
         else:
-            return HttpResponseBadRequest(f"User Exists Or Form Have Invalid Failed/s.")
+            data_set = dict(detail="User Exists Or Form Have Invalid Failed/s.")
+            json_response = json.dumps(data_set)
+            return HttpResponseBadRequest(json_response)
     else:
-        return HttpResponseBadRequest("Bad request.")
+        data_set = dict(detail="Bad request.")
+        json_response = json.dumps(data_set)
+        return HttpResponseBadRequest(json_response)
 
 
+@login_required
 @csrf_exempt
 def like_post(request):
     if request.method == 'POST':
@@ -80,13 +100,21 @@ def like_post(request):
             post_ = models.Post.objects.get(pk=post_id)
             user_ = models.User.objects.get(username=user)
             post_.post_likes.add(user_)
-            return JsonResponse({"User": user_.username, "liked Post": post_.post_content})
+            data_set = dict(user=user_.username, post={
+                "post_id": post_.id,
+                "author": post_.author.username,
+                "post_content": post_.post_content})
+            json_response = json.dumps(data_set)
+            return HttpResponse(json_response)
         else:
-            return HttpResponseNotFound("User or Post does not exist.")
+            json_response = json.dumps(dict(detail="User or Post does not exist."))
+            return HttpResponseNotFound(json_response)
     else:
-        return HttpResponseBadRequest("Bad request.")
+        json_response = json.dumps(dict(detail="Bad request."))
+        return HttpResponseBadRequest(json_response)
 
 
+@login_required
 @csrf_exempt
 def share_post(request):
     if request.method == 'POST':
@@ -100,14 +128,25 @@ def share_post(request):
             post_.post_shares.add(user_)
             new_shared_post = models.Post(author=user_, post_content=post_.post_content)
             new_shared_post.save()
-            return HttpResponse(f"User: {user_.username} shared Post: {post_.post_content}"
-                                f"\nNew post created.")
+            data_set = dict(user=user_.username,
+                            shared_post={
+                                "post_id": post_.id,
+                                "author": post_.author.username,
+                                "post_content": post_.post_content},
+                            new_post={
+                                "post_id": new_shared_post.id
+                            })
+            json_response = json.dumps(data_set)
+            return HttpResponse(json_response)
         else:
-            return HttpResponseNotFound("User or Post does not exist.")
+            json_response = json.dumps(dict(detail="User or Post does not exist."))
+            return HttpResponseNotFound(json_response)
     else:
-        return HttpResponseBadRequest("Bad request!")
+        json_response = json.dumps(dict(detail="Bad request."))
+        return HttpResponseBadRequest(json_response)
 
 
+@login_required
 @csrf_exempt
 def unlike_post(request):
     if request.method == 'POST':
@@ -119,13 +158,21 @@ def unlike_post(request):
             post_ = models.Post.objects.get(pk=post_id)
             user_ = models.User.objects.get(username=user)
             post_.post_likes.remove(user_)
-            return HttpResponse(f"User: {user_.username} unliked Post: {post_.post_content}")
+            data_set = dict(user=user_.username, post={
+                "post_id": post_.id,
+                "author": post_.author.username,
+                "post_content": post_.post_content})
+            json_response = json.dumps(data_set)
+            return HttpResponse(json_response)
         else:
-            return HttpResponseNotFound("User or Post does not exist.")
+            json_response = json.dumps(dict(detail="User or Post does not exist."))
+            return HttpResponseNotFound(json_response)
     else:
-        return HttpResponseBadRequest("Bad request!")
+        json_response = json.dumps(dict(detail="Bad request."))
+        return HttpResponseBadRequest(json_response)
 
 
+@login_required
 @csrf_exempt
 def post_post(request):
     if request.method == 'POST':
@@ -136,55 +183,81 @@ def post_post(request):
             user_ = models.User.objects.get(username=user)
             new_post = models.Post(author=user_, post_content=post_content)
             new_post.save()
-            return HttpResponse(f"New blog posted: {post_content}, Author: {user_.username}")
+            data_set = dict(
+                post_id=new_post.id,
+                author=new_post.author.username,
+                post_content=new_post.post_content)
+            json_response = json.dumps(data_set)
+            return HttpResponse(json_response)
         else:
-            return HttpResponseNotFound("User does not exist.")
+            json_response = json.dumps(dict(detail="User does not exist."))
+            return HttpResponseNotFound(json_response)
     else:
-        return HttpResponseBadRequest("Bad request.")
+        json_response = json.dumps(dict(detail="Bad request."))
+        return HttpResponseBadRequest(json_response)
 
 
+@login_required
 @csrf_exempt
 def edit_post(request):
     if request.method == 'POST':
         user = request.user
         post_new_content = request.POST.get('new_content')
-        post = request.POST.get('post_id')
+        post_id = request.POST.get('post_id')
+        post_exist = models.Post.objects.filter(id=post_id).count()
         user_exist = models.User.objects.filter(username=user).count()
-        post_ = models.Post.objects.get(pk=post)
-        old_content = post_.post_content
-        if user_exist > 0:
+        if user_exist > 0 and post_exist > 0:
+            post_ = models.Post.objects.get(pk=post_id)
             if user.id == post_.author.id:
+                old_content = post_.post_content
                 post_.post_content = post_new_content
                 post_.save()
-                return HttpResponse(f"Blog edit: \nNew content - {post_new_content},\n"
-                                    f"Old content - {old_content}")
+                data_set = dict(
+                    post_id=post_.id,
+                    author=post_.author.username,
+                    new_post_content=post_.post_content,
+                    old_post_content=old_content)
+                json_response = json.dumps(data_set)
+                return HttpResponse(json_response)
             else:
-                return HttpResponseNotAllowed(f"The user {user} can not edit this post.")
+                json_response = json.dumps(dict(detail=f"The user {user} can not edit this post."))
+                return HttpResponseNotAllowed(json_response)
         else:
-            return HttpResponseNotFound("User does not exist.")
+            json_response = json.dumps(dict(detail="User Or Post does not exist."))
+            return HttpResponseNotFound(json_response)
     else:
-        return HttpResponseBadRequest("Bad request.")
+        json_response = json.dumps(dict(detail="Bad request."))
+        return HttpResponseBadRequest(json_response)
 
 
+@login_required
 @csrf_exempt
 def delete_post(request):
     if request.method == 'POST':
         user = request.user
-        post = request.POST.get('post_id')
+        post_id = request.POST.get('post_id')
+        post_exist = models.Post.objects.filter(id=post_id).count()
         user_exist = models.User.objects.filter(username=user).count()
-        post_ = models.Post.objects.get(pk=post)
-        if user_exist > 0:
-            if post_:
-                if user.id == post_.author.id:
-                    post_content = post_.post_content
-                    post_.delete()
-                    return HttpResponse(f"Post '{post_content}' deleted")
-                else:
-                    return HttpResponseNotAllowed(f"The user {user} cannot delete this post."
-                                                  f"\nThe author just can delete it.")
+        if user_exist > 0 and post_exist > 0:
+            post_ = models.Post.objects.get(pk=post_id)
+            if user.id == post_.author.id:
+                post_id = post_.id
+                author = post_.author.username
+                post_content = post_.post_content
+                post_.delete()
+                data_set = dict(
+                    post_id=post_id,
+                    author=author,
+                    post_content=post_content)
+                json_response = json.dumps(data_set)
+                return HttpResponse(json_response)
             else:
-                HttpResponseNotFound("Post does not exist.")
+                data_set = dict(response=f"The user {user} is not the post's author.")
+                json_response = json.dumps(data_set)
+                return HttpResponseBadRequest(json_response, status=401)
         else:
-            HttpResponseNotFound("User does not exist.")
+            json_response = json.dumps(dict(detail="Post or User does not exist."))
+            return HttpResponseNotFound(json_response)
     else:
-        return HttpResponseBadRequest("Bad request!")
+        json_response = json.dumps(dict(detail="Bad request!"))
+        return HttpResponseBadRequest(json_response)
